+++ lib/paper_trail/record_trail.rb
@@ -148,11 +148,14 @@ module PaperTrail
     # paper_trail-association_tracking
     def record_update_columns(changes)
       return unless enabled?
-      event = Events::Update.new(@record, false, false, changes)
+      data = Events::Update.new(@record, false, false, changes).data
+
+      # use AR default timestamp value instead of record.updated_at value
+      data.delete(:created_at)
 
       # Merge data from `Event` with data from PT-AT. We no longer use
       # `data_for_update_columns` but PT-AT still does.
-      data = event.data.merge(data_for_update_columns)
+      data.merge!(data_for_update_columns)
 
       versions_assoc = @record.send(@record.class.versions_association_name)
       version = versions_assoc.create(data)