+++ spec/spec_helper.rb
@@ -10,6 +10,7 @@ end
 SimpleCov.minimum_coverage(ENV["DB"] == "postgres" ? 97.3 : 92.4)
 
 require "byebug"
+require "active_support/testing/time_helpers"
 require_relative "support/pt_arel_helpers"
 
 unless ENV["BUNDLE_GEMFILE"].match?(/rails_\d\.\d\.gemfile/)
@@ -41,6 +42,7 @@ RSpec.configure do |config|
   end
   config.order = :random
   config.include PTArelHelpers
+  config.include ActiveSupport::Testing::TimeHelpers
   Kernel.srand config.seed
 end